Implement new list structure
This commit introduces a new list structure, which is more performant
and has a cleaner implementation than the previous. It's base is a
circular doubly linked list, which has O(1) on append, prepend, insert,
start, end, next and previous. This means that this structure is the
fastest possible linked list.

The list was designed to provide if-less operations, which means that
the code will never generate conditional jumps. This is especially good
for x86-like CPUs, where on a conditional jump the pipeline must be
flushed before proceeding.

Reminder: Don't use this structure if you need O(1) access to random
elements, this will perform in worst case in O(n). Alternatives are a
binary search tree O(log(n)) in average or an array O(1) in average and
worst case.